Add preview-template plugin
