+++ packages/@docfy/ember/dummy-docs/packages/ember/index.md
@@ -1,3 +1,22 @@
 # Working with Ember
 
 This is a simple markdown
+
+
+Where are a lot of cool features here. One of them is to preview a template.
+
+```hbs preview-template
+<DocfyLink @to="/" class="text-blue-700">This is a link to index</DocfyLink>
+
+<div class="p-4 my-2 border border-blue-500 rounded">
+  I can have any hbs here and I can use any component avaliable in the host app.
+</div>
+```
+
+## Another content
+
+Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
+
+```hbs preview-template
+Another preview here
+```
